reorganized the server folders
