+++ server/routes.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const controllers = require('./controllers.js');
+const controllers = require('./MVC/controllers.js');
 
 router.get('/:product_id', controllers.getProduct);
 router.get('/:product_id/styles', controllers.getProductStyles);
